refactor(server): split api.rs with regression tests pinning security constants (#772)
* refactor(server): split api.rs into sessions / git / system submodules

Splits src/server/api.rs (2,151 lines, 28 axum handlers + DTOs) into:
  - src/server/api/mod.rs (50 lines) — module declarations, re-exports of
    every public handler so the route table in src/server/mod.rs is
    unchanged, and the shared validation helpers.
  - src/server/api/sessions.rs (1,484 lines) — session CRUD, ensure-*
    lifecycle endpoints, and per-file rich-diff handlers.
  - src/server/api/git.rs (293 lines) — clone_repo, list_branches,
    repo_name_from_url + tests.
  - src/server/api/system.rs (384 lines) — agents, settings, themes,
    profiles, filesystem browse, groups, docker status, devices, about.

While splitting:
  - Security constants (SHELL_METACHARACTERS, ALLOWED_SETTINGS_SECTIONS,
    SESSION_BLOCKED_FIELDS) are carried over byte-for-byte. A regression
    test that pins them lives in the follow-up commit on this branch.
  - browse_filesystem: sort_by(|a,b| to_lowercase.cmp) -> sort_by_cached_key
    (O(n) alloc instead of O(n log n)).
  - list_branches: sort_by(|a,b| cmp) -> sort_by_key + Reverse.
  - Doc-list indent in api/mod.rs normalized so rustdoc's overindent lint
    stays clean.

1,213 lib tests + every integration suite pass.


* test(server): pin security constants against silent regression

Add cargo-test regressions that fail the build if any of the three
security-critical constants in src/server/api/mod.rs are silently
changed during a refactor.

Motivation: an earlier hand-assembled version of the api.rs split
rewrote these constants by hand instead of carrying them over, and
all three regressed:
  - SHELL_METACHARACTERS dropped `#`, `[`, `]`, `~` (shell-injection
    vectors: comment, glob, tilde expansion).
  - ALLOWED_SETTINGS_SECTIONS gained 'hooks' (API writes to hooks
    bypass the repo-hook trust prompt and run arbitrary commands on
    session start — local RCE).
  - SESSION_BLOCKED_FIELDS was replaced with two field names that
    don't exist on SessionConfig, silently turning the blocklist
    off.

Four tests pin each list exhaustively by size-and-content and assert
that validate_no_shell_injection rejects every listed character. A
size mismatch now fails CI with a message explaining the security
context and forcing the author to justify the change instead of
smuggling it through a 'no behavior changes' refactor commit.

Also picks up cargo fmt normalization of a few use statements in the
split submodules.


---------
